requirements.txt
asyncio==3.4.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kucoin-futures-python==1.0.6
kucoin-python==1.0.11
MarkupSafe==2.1.1
requests==2.28.1
urllib3==1.26.10
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.1
asyncio==3.4.3
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kucoin-futures-python==1.0.6
kucoin-python==1.0.11
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
pandas==1.5.0
pandas-ta==0.3.14b0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-orm==1.2.3
typing-extensions==4.3.0
typing-inspect==0.8.0
urllib3==1.26.10
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.1
alembic==1.9.1
asyncio==3.4.3
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.10.2
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kucoin-futures-python==1.0.6
kucoin-python==1.0.11
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
pandas==1.5.0
pandas-ta==0.3.14b0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-orm==1.2.3
typing-extensions==4.3.0
typing-inspect==0.8.0
urllib3==1.26.10
websockets==10.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
